Moved load balancer functions out into a provider file
