+++ dbt_platform_helper/commands/environment.py
@@ -11,6 +11,7 @@ from schema import SchemaError
 
 from dbt_platform_helper.constants import DEFAULT_TERRAFORM_PLATFORM_MODULES_VERSION
 from dbt_platform_helper.constants import PLATFORM_CONFIG_FILE
+from dbt_platform_helper.providers.load_balancers import find_https_listener
 from dbt_platform_helper.utils.application import Environment
 from dbt_platform_helper.utils.application import Service
 from dbt_platform_helper.utils.application import load_application
@@ -369,44 +370,6 @@ def _determine_terraform_platform_modules_version(env_conf, cli_terraform_platfo
         DEFAULT_TERRAFORM_PLATFORM_MODULES_VERSION,
     ]
     return [version for version in version_preference_order if version][0]
-
-
-def find_load_balancer(session: boto3.Session, app: str, env: str) -> str:
-    lb_client = session.client("elbv2")
-
-    describe_response = lb_client.describe_load_balancers()
-    load_balancers = [lb["LoadBalancerArn"] for lb in describe_response["LoadBalancers"]]
-
-    load_balancers = lb_client.describe_tags(ResourceArns=load_balancers)["TagDescriptions"]
-
-    load_balancer_arn = None
-    for lb in load_balancers:
-        tags = {t["Key"]: t["Value"] for t in lb["Tags"]}
-        if tags.get("copilot-application") == app and tags.get("copilot-environment") == env:
-            load_balancer_arn = lb["ResourceArn"]
-
-    if not load_balancer_arn:
-        raise LoadBalancerNotFoundError()
-
-    return load_balancer_arn
-
-
-def find_https_listener(session: boto3.Session, app: str, env: str) -> str:
-    load_balancer_arn = find_load_balancer(session, app, env)
-    lb_client = session.client("elbv2")
-    listeners = lb_client.describe_listeners(LoadBalancerArn=load_balancer_arn)["Listeners"]
-
-    listener_arn = None
-
-    try:
-        listener_arn = next(l["ListenerArn"] for l in listeners if l["Protocol"] == "HTTPS")
-    except StopIteration:
-        pass
-
-    if not listener_arn:
-        raise ListenerNotFoundError()
-
-    return listener_arn
 
 
 def find_https_certificate(session: boto3.Session, app: str, env: str) -> str:
